+++ src/main/java/challenging/application/exception/challenge/InvalidDateException.java
@@ -1,4 +1,4 @@
-package challenging.application.exeption.challenge;
+package challenging.application.exception.challenge;
 
 public class InvalidDateException extends RuntimeException {
